+++ packages/tinacms/src/toolkit/fields/plugins/mdx-field-plugin/plate/components/floating-toolbar-buttons.tsx
@@ -1,7 +1,7 @@
 import React from 'react';
 import { TurnIntoDropdownMenu } from './plate-ui/turn-into-dropdown-menu';
-import { CommentToolbarButton } from '../../../streamliners/discussion-plugin/comment-toolbar-button';
-import { SuggestionToolbarButton } from '../../../streamliners/suggestion-plugin/suggestion-toolbar-button';
+import { CommentToolbarButton } from '../../../streamliners/discussion-plugin/components/comment-toolbar-button';
+import { SuggestionToolbarButton } from '../../../streamliners/suggestion-plugin/components/suggestion-toolbar-button';
 
 export default function FloatingToolbarButtons() {
   return (
