+++ packages/tinacms/src/toolkit/fields/plugins/streamliners/discussion-plugin/components/block-thread-view.tsx
@@ -1,55 +1,48 @@
 'use client';
 
 import React from 'react';
-import { useEditorPlugin, usePluginOption } from '@udecode/plate/react';
-import { useCMS } from '@toolkit/react-core';
+import { useEditorPlugin } from '@udecode/plate/react';
 
-import { CommentMessage, CommentInput, SuggestionDiff } from './shared';
+import {
+  CommentInput,
+  SuggestionDiff,
+  ThreadMessages,
+  ThreadSubject,
+} from '.';
 import {
   commentPlugin,
   type CommentThread,
-} from './comment-plugin';
-import { suggestionPlugin } from '../suggestion-plugin/suggestion-plugin';
+} from '../plugins/comment-plugin';
+import { suggestionPlugin } from '../../suggestion-plugin/suggestion-plugin';
 import {
   appendMessageToThread,
   getSuggestionDiff,
   acceptActiveSuggestion,
   rejectActiveSuggestion,
   clearCommentThread,
-} from './annotation-util';
-import type { StoredSuggestion } from './annotations-store';
+} from '../utils/annotation-util';
+import {
+  useAnnotationThreads,
+  useAnnotationUser,
+} from '../hooks/use-annotation-state';
 
 interface BlockThreadViewProps {
   threadId: string;
   isSuggestion?: boolean;
 }
 
-type CurrentUser = { id?: string; name?: string } | null;
-
 export function BlockThreadView({ threadId, isSuggestion = false }: BlockThreadViewProps) {
-  const cms = useCMS();
   const { editor, setOption: setCommentOption } = useEditorPlugin(commentPlugin);
   const { setOption: setSuggestionOption } = useEditorPlugin(suggestionPlugin);
+  const currentUser = useAnnotationUser();
+  const { getThreads, commitThread: commitThreadUpdate, deleteThread } =
+    useAnnotationThreads();
 
   // Get comments directly from plugin options (single source of truth)
-  const comments = (editor.getOption(commentPlugin, 'threads') || {}) as Record<string, CommentThread>;
-
-  const [currentUser, setCurrentUser] = React.useState<CurrentUser>(null);
+  const comments = getThreads();
   const [replyValue, setReplyValue] = React.useState('');
   const [editingMessageId, setEditingMessageId] = React.useState<string | null>(null);
   const [editingValue, setEditingValue] = React.useState('');
-
-  React.useEffect(() => {
-    void cms.api.tina.authProvider.getUser().then((user) => {
-      if (!user) {
-        setCurrentUser({ id: 'anonymous', name: 'Anonymous' });
-        return;
-      }
-      const id = user.id ?? user.email ?? 'anonymous';
-      const name = user.name ?? user.email ?? 'Anonymous';
-      setCurrentUser({ id, name });
-    });
-  }, [cms.api.tina.authProvider]);
 
   const thread: CommentThread | null = comments[threadId] ?? null;
 
@@ -57,17 +50,6 @@ export function BlockThreadView({ threadId, isSuggestion = false }: BlockThreadV
     if (!isSuggestion) return null;
     return getSuggestionDiff(editor, threadId);
   }, [editor, threadId, isSuggestion]);
-
-  const commitThreadUpdate = React.useCallback(
-    (nextThread: CommentThread) => {
-      const currentThreads = editor.getOption(commentPlugin, 'threads') || {};
-      editor.setOption(commentPlugin, 'threads', {
-        ...currentThreads,
-        [nextThread.id]: nextThread,
-      });
-    },
-    [editor]
-  );
 
   const handleReplySubmit = () => {
     const value = replyValue.trim();
@@ -159,11 +141,7 @@ export function BlockThreadView({ threadId, isSuggestion = false }: BlockThreadV
       userId: currentUser?.id,
     });
 
-    const currentThreads = editor.getOption(commentPlugin, 'threads') || {};
-    if (currentThreads[threadId]) {
-      const { [threadId]: _removed, ...rest } = currentThreads;
-      editor.setOption(commentPlugin, 'threads', rest);
-    }
+    deleteThread(threadId);
     if (isSuggestion && suggestionDiff) {
       clearCommentThread(editor, threadId);
       setSuggestionOption('activeId', null);
@@ -180,26 +158,14 @@ export function BlockThreadView({ threadId, isSuggestion = false }: BlockThreadV
       userId: currentUser?.id,
     });
 
-    const currentThreads = editor.getOption(commentPlugin, 'threads') || {};
-    if (currentThreads[threadId]) {
-      const { [threadId]: _removed, ...rest } = currentThreads;
-      editor.setOption(commentPlugin, 'threads', rest);
-    }
+    deleteThread(threadId);
     clearCommentThread(editor, threadId);
     setSuggestionOption('activeId', null);
     setCommentOption('activeId', null);
   };
 
   const handleDeleteThread = () => {
-    editor.getTransforms(commentPlugin).comment.unsetMark?.({
-      id: threadId,
-    });
-
-    const currentThreads = editor.getOption(commentPlugin, 'threads') || {};
-    if (currentThreads[threadId]) {
-      const { [threadId]: _removed, ...rest } = currentThreads;
-      editor.setOption(commentPlugin, 'threads', rest);
-    }
+    deleteThread(threadId);
 
     if (isSuggestion && suggestionDiff) {
       clearCommentThread(editor, threadId);
@@ -219,50 +185,22 @@ export function BlockThreadView({ threadId, isSuggestion = false }: BlockThreadV
         />
       )}
 
-      {thread?.discussionSubject && (
-        <div className="flex flex-col gap-1.5">
-          <div className="text-sm font-medium text-foreground">Subject</div>
-          <blockquote className="rounded-md border-l-4 border-blue-500 bg-blue-500/10 p-3 text-sm text-foreground">
-            {thread.discussionSubject}
-          </blockquote>
-        </div>
-      )}
+      <ThreadSubject subject={thread?.discussionSubject} />
 
-      {thread?.messages?.length ? (
-        <div className="flex flex-col gap-1.5">
-          <div className="text-sm font-medium">Discussion</div>
-          <div className="max-h-[280px] overflow-y-auto">
-            <div className="flex flex-col">
-              {(thread.messages ?? []).map((message, index) => {
-                const isLast = index === (thread.messages ?? []).length - 1;
-                return (
-                  <CommentMessage
-                    key={message.id}
-                    id={message.id}
-                    body={message.body}
-                    authorId={message.authorId}
-                    authorName={message.authorName}
-                    createdAt={message.createdAt}
-                    updatedAt={message.updatedAt}
-                    currentUserId={currentUser?.id}
-                    isLast={isLast}
-                    onEdit={handleEditMessage}
-                    onDelete={handleDeleteMessage}
-                    isEditing={editingMessageId === message.id}
-                    editingValue={editingValue}
-                    onEditingValueChange={setEditingValue}
-                    onSubmitEdit={handleSubmitEdit}
-                    onCancelEdit={() => {
-                      setEditingMessageId(null);
-                      setEditingValue('');
-                    }}
-                  />
-                );
-              })}
-            </div>
-          </div>
-        </div>
-      ) : null}
+      <ThreadMessages
+        messages={thread?.messages}
+        currentUserId={currentUser?.id}
+        editingMessageId={editingMessageId}
+        editingValue={editingValue}
+        onEdit={handleEditMessage}
+        onDelete={handleDeleteMessage}
+        onEditingValueChange={setEditingValue}
+        onSubmitEdit={handleSubmitEdit}
+        onCancelEdit={() => {
+          setEditingMessageId(null);
+          setEditingValue('');
+        }}
+      />
 
       <CommentInput
         value={replyValue}
