+++ packages/tinacms/src/toolkit/fields/plugins/streamliners/discussion-plugin/components/index.ts
@@ -1,0 +1,8 @@
+export { CommentMessage, type CommentMessageProps } from './comment-message';
+export { SuggestionDiff, type SuggestionDiffProps } from '../../suggestion-plugin/components/suggestion-diff';
+export { CommentInput, type CommentInputProps } from './comment-input';
+export {
+  ThreadSubject,
+  ThreadMessages,
+  type ThreadMessagesProps,
+} from './thread';
