+++ packages/tinacms/src/toolkit/fields/plugins/streamliners/discussion-plugin/index.ts
@@ -1,13 +1,15 @@
-export * from './comment-plugin';
-// export * from './discussion-popover';
-export * from './comment-state-synchronizer';
-export * from './comment-toolbar-button';
-export * from './comment-node';
-export * from './annotation-popover';
+export * from './plugins/comment-plugin';
+export * from './components/comment-toolbar-button';
+export * from './components/comment-node';
+export * from './components/annotation-popover';
 export * from './types';
-export * from './discussion-adapter';
-export * from './discussion-plugin';
-export * from './discussion-plugin-bridge';
-export * from './annotations-store';
-export * from './annotation-plugin-bridge';
-export * from './comment-annotations';
+export * from './plugins/discussion-adapter';
+export * from './plugins/discussion-plugin';
+export * from './utils/annotations-store';
+export * from './utils/comment-annotations';
+export * from '../suggestion-plugin/suggestion-plugin';
+export * from '../suggestion-plugin/components/suggestion-leaf';
+export * from '../suggestion-plugin/components/suggestion-toolbar-button';
+export * from '../suggestion-plugin/components/suggestion-card';
+export * from '../suggestion-plugin/utils/suggestion-state-synchronizer';
+export * from '../suggestion-plugin/utils/suggestion-helpers';
