--- packages/tinacms/src/toolkit/fields/plugins/streamliners/discussion-plugin/shared/index.ts
@@ -1,3 +1,0 @@
-export { CommentMessage, type CommentMessageProps } from './CommentMessage';
-export { SuggestionDiff, type SuggestionDiffProps } from './SuggestionDiff';
-export { CommentInput, type CommentInputProps } from './CommentInput';
