+++ packages/tinacms/src/toolkit/fields/plugins/streamliners/discussion-plugin/utils/annotation-util.ts
@@ -1,0 +1,13 @@
+export { createAnnotationId } from './annotation-id';
+export { appendMessageToThread } from './thread-helpers';
+export {
+  ensureCommentMark,
+  clearCommentThread,
+  DRAFT_COMMENT_KEY,
+} from './comment-marks';
+export {
+  getSuggestionDiff,
+  buildResolvedSuggestion,
+  acceptActiveSuggestion,
+  rejectActiveSuggestion,
+} from '../../suggestion-plugin/utils/suggestion-helpers';
