+++ packages/tinacms/src/toolkit/fields/plugins/streamliners/suggestion-plugin/index.ts
@@ -1,5 +1,4 @@
 export * from './suggestion-plugin';
-export * from './suggestion-leaf';
-export * from './suggestion-toolbar-button';
-export * from './suggestion-popover';
-export * from './suggestion-state-synchronizer';
+export * from './components/suggestion-leaf';
+export * from './components/suggestion-toolbar-button';
+export * from './utils/suggestion-state-synchronizer';
