--- packages/tinacms/src/toolkit/fields/plugins/streamliners/suggestion-plugin/suggestion-popover.tsx
@@ -1,14 +1,0 @@
-'use client';
-
-import React from 'react';
-
-import { AnnotationPopover } from '../discussion-plugin/annotation-popover';
-
-/**
- * @deprecated Replaced by `AnnotationPopover`. This component now renders the unified popover.
- */
-export function SuggestionPopover() {
-  return <AnnotationPopover />;
-}
-
-export default SuggestionPopover;
